Refactored AATA functions into extracters.aata_data. Add identifiers for abstracts.
